+++ src/main/java/com/academy/course/appcafe/dto/ProductDTO.java
@@ -5,16 +5,18 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Builder
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductDTO {
-    private Integer id;
+    private Long id;
 
     private String name;
 
-    private Double price;
+    private BigDecimal price;
 
     private String info;
